+++ packages/colony-js-contract-loader/src/interface/ContractLoader.js
@@ -6,6 +6,7 @@ export type Query = {
   routerAddress?: string,
   routerName?: string,
   version?: string,
+  networkId?: string,
 };
 
 export type ContractDefinition = {