+++ packages/colony-js-contract-loader/src/transforms/truffleTransform.js
@@ -1,24 +1,33 @@
 /* @flow */
 
+import type { Query } from '../interface/ContractLoader';
+
 type TruffleArtifact = {
   abi: Array<{}>,
   bytecode: string,
   networks: {
-    [networkId: number]: {
+    [networkId: string | number]: {
       address: string,
     },
   },
 };
 
-export default function truffleTransform({
-  abi = [],
-  bytecode,
-  networks = {},
-}: TruffleArtifact = {}) {
+export default function truffleTransform(
+  { abi = [], bytecode, networks = {} }: TruffleArtifact = {},
+  { networkId }: Query = {},
+) {
+  let address;
+
   const networkIds = Object.keys(networks);
-  // Pick the last network (assumed to be the most recent)
-  const { address } =
-    networks[parseInt(networkIds[networkIds.length - 1], 10)] || {};
+
+  if (networkId && networkIds.length) {
+    if (!networks[networkId])
+      throw new Error(`Network ID ${networkId} not found in contract`);
+    ({ address } = networks[networkId]);
+  } else {
+    // Pick the last network (assumed to be the most recent)
+    ({ address } = networks[networkIds[networkIds.length - 1]] || {});
+  }
 
   return {
     abi,